+++ webextension/data/js/classes/ZDK.js
@@ -287,6 +287,38 @@ class ZDK{
 		}
 	}
 
+	static async openTabIfNotExist(url){
+		consoleMsg("log", url);
+
+		const tabs = await browser.tabs.query({});
+
+		let custom_url = url.toLowerCase().replace(/http(?:s)?:\/\/(?:www\.)?/i,"");
+		for(let tab of tabs){
+			if(tab.url.toLowerCase().indexOf(custom_url) !== -1){ // Mean the url was already opened in a tab
+				browser.tabs.highlight({tabs: tab.index}); // Show the already opened tab
+				return true; // Return true to stop the function as the tab is already opened
+			}
+		}
+
+		const browserWindows = await browser.windows.getAll({
+			populate: false,
+			windowTypes: ["normal"]
+		});
+
+		// If the function is still running, it mean that the url isn't detected to be opened, so, we can open it
+		if(browserWindows.length===0){
+			await browser.windows.create({
+				"focused": true,
+				"type": "normal",
+				"url": url
+			});
+		} else{
+			await browser.tabs.create({ "url": url });
+		}
+
+		return false; // Return false because the url wasn't already in a tab
+	}
+
 	/**
 	 *
 	 * @param action